refactor wrapped instructions to lower duplication
This refactor extracts the read/write methods along with
the size getter to a parent class `InstructionWrapper`.
All of the classes that now subclass `InstructionWrapper`
can override their parent method to change the behavior
of the above methods/getter if they need to do so by
implementing the relevant override method they need.